Fully remove networking I/O from core. A network connection is now identified by three things     int socket, an identifier for the transport, never dereferenced     uint8_t *addr + size_t addrLen, an opaque buffer used by the the application to identify a destination
Socket transmit and receive code now lives purely in the application, with a callback being made when the core wishes to send data (see lwm2m_init()).

The current client and server apps are largely unchanged.
Future applications should be able to use the core without a BSD sockets interface.