bootstrap, first step implementation
You can add a bootstrap server, and trigger a bootstrap.
A 'bootstrap' command is added to the client.

Bootstrap doesn't use a transaction because currently transaction.c
works only with lwm2m_server now.
Bootstrap won't work now because the bootstrap server will try to DELETE
the whole /1/ hierachy and provision new servers which is cleary not
supported.

Signed-off-by: Julien Vermillard <jvermillard@sierrawireless.com>