Add a bootstrap timeout to fix issue #282
er-coap-13.h is modified to calculate the exchange lifetime.

When the ACK to the bootstrap request, or any bootstrap command is
received, the server registration time is updated to be the current
time plus the exchange lifetime. If this time has passed in the
pending state, the state is set to failing.

Signed-off-by: Scott Bertin <sbertin@telular.com>