Cancel the correct observation.
When a cancel observation was received it would always cancel the most
recent observation rather than the one specified in the message.

Steps to reproduce:
Build client with LWM2M_WITH_LOGS.
Connect to server.
On server observe /31024/10 and /31024/11 in that order.
On server cancel /31024/10.
Client observe_step logs show /31024/10 and not /31024/11 indicating the
wrong observe was canceled.

This change looks up the correct mid to cancel and requires the correct
mid be passed to observe_cancel. In addition, the last mid is set to the
initial response mid to allow correct identification of the observation
when there have been no notifies.

Signed-off-by: Scott Bertin <sbertin@telular.com>