Fix cancel observation with multiple watchers.
With multiple servers an observe cancellation may not take effect.

Steps to reproduce:
Connect the client to multiple servers.
On server 1, observe /31024/10.
On server 2, observe /31024/10.
On server 1 cancel /31024/10.
On the client change /31024/10/1.
Both servers are still notified.

This change corrects the check of the session. It was skipping over in
case of a match and should skip in case of a mismatch.

Signed-off-by: Scott Bertin <sbertin@telular.com>