Make TLV format optional.
It is always enabled for LWM2M 1.0 clients and servers and LWM2M 1.1
servers, but disabled by default for LWM2M 1.1 clients.

Observe cancel responses from the client and Create requests from the
server were hard coded to use TLV.

Observe cancel will now respect the format in the accept option or use
the preferred format of the client if there is no accept option.

Create requests will now use the preferred client format except when no
object instance is specified. JSON and SenML JSON formats do not parse
or serialize properly with a missing object instance, so TLV is forced
in that case until it is fixed.

Reading an object instance that contained only a single resource could
respond with text/plain content, which is only valid when reading the
resource, not the instance. This adds more knowledge of what was read to
the function determining the response format.

Signed-off-by: Scott Bertin <sbertin@telular.com>