Removed forced usage of TLV format.
Observe cancel responses from the client and Create requests from the
server were hard coded to use TLV.

Observe cancel will now respect the format in the accept option or use
the preferred format of the client if there is no accept option.

Create requests will now use the preferred client format except when no
object instance is specified. JSON and SenML JSON formats do not parse
or serialize properly with a missing object instance, so TLV is forced
in that case until it is fixed.


Signed-off-by: Scott Bertin <sbertin@telular.com>