Implement lwm2m_gettime() using time().
Currently lwm2m_gettime() is implemented using POSIX's gettimeofday().
This creates a number of issues:

1. gettimeofday() gives the "time of day", meaning it overflows each
   24 hours. This runs contrary to the specification in liblwm2m.h,
   an may result in wring behaviour.
2. gettimeofday() is marked obsolete in POSIX 2008.
3. time is standard in C89/C99, not only POSIX.
4. Using time() is more straightforward.

I understand that gettimeofday() used only in the examples and it not
part of the waakama library. The examples, however, are what people
wanting use waakama will look at first and probably copy, so they should
be correct. Using gettimeofday() creates confusion as to whether the
absolute time or the time since 0AM should be used.

The first issue is the most important. Looking at the code, there is no
evidence that the 24-hour overflow is being taken into account. So, by
default we must assime that an overflowing time value may break waakama.

For example, in core/transaction.c:389 we find:
```c
transacP->retrans_time = tv_sec + COAP_RESPONSE_TIMEOUT;
```

and later in core/transaction.c:443 :
```c
if (transacP->retrans_time <= currentTime)
```

If tv_sec is acquired via gettimeofday() it has a range from 0 to 3600*24,
and the first line may put `transacP->retrans_time` outside that range. If
that happens, the comparison in line 443 will always be false.

Point (3) is relevant for those trying to use waakama in embedded system
where the full POSIX API is not available. Using time() makes it clear that
waakama does not need POSIX time functions. Now, I'm not very familiar with
this code, but it seems to me that the library itself does not need POSIX
_at all_.

Point (2) and (4) are easy to verify.

I haven't tested this, apart from checking that it stil compiles. I do not
know how to test waakama.

Signed-off-by: Juan Carrano <j.carrano@fu-berlin.de>