Add context to all callbacks
Some activities require knowing the LWM2M context and may be triggered
by callbacks. Examples of this are server object execute resources to
trigger a registration update or bootstrap. Before this change it is
necessary to have a means outside Wakaama to get that context. This
provides the context to all callbacks so that there is uniform access to
the context.

Signed-off-by: Scott Bertin <sbertin@telular.com>