Bootstrap Delete /0 keep bootstrap server account
The LWM2M spec states:
Only in Bootstrap Interface, the Delete operation MAY targets “/0” URI.
In that case, the operation MUST delete all the existing Security
Object Instances - except LwM2M Bootstrap-Server Account - in the LwM2M
Client;

This changes the logic for the Delete /0 case to match the Delete All.
It avoids deleting the Security object for the bootstrap server.

Signed-off-by: Scott Bertin <sbertin@telular.com>