fix: copy transaction payload to prevent use after free
The caller of a method (e.g. lwm2m_dm_write) could have free'd the buffer after
the method completes (assuming that he retains ownership of allocated memory).
But the payload buffer is used by wakaama even beyond the scope of the method /
current transaction. This happens for block transfers but could also happen in
case of re-transmissions. For such cases, it is necessary for wakaama to copy
the payload for the transaction.

Signed-off-by: Pascal Brogle <pascal.brogle@husqvarnagroup.com>