Fix LWM2M_RAW_BLOCK1_REQUESTS compile
Addresses the compile issues referenced in #773 when using the
`LWM2M_RAW_BLOCK1_REQUESTS` compile flag by:
- removing unused variables
- specifying the missing const keyword.

In addition, to help validate that this doesn't happen in the future, this
commit adds an additional configuration option to the client example which
enables the `LWM2M_RAW_BLOCK1_REQUESTS` option. This allows for a basic test
that compilation succeeds. It does not go further in attempting to add unit
tests for this compile option.

Signed-off-by: David Smith <mmb-davidsmith@users.noreply.github.com>