cmake: Allow setting mode with CMake
The Wakaama modes (server, bootstrap server or client) can now be set
with CMake cache variables. Some CMake presets are provided to build,
run and test Wakaama with different mode configurations.
The example binaries that are needed for the integration tests are now
built with the CMake presets. Therefore the paths of the binaries were
adjusted in the integration tests.