cmake: Extract connection for testing in library
The testing implementation of the transport layer is extracted in an own
object library and linked directly to the unit tests.
Also other needed libraries (CLI and platform) are directly linked to
the testing executable.
The possibility to select the testing transport layer from CMake is
removed. This special transport layer implementation is meant to be used
only in the unit tests internally.:X