cmake: Add Wakaama static library
The static library can be configured with CMake variables and then be
linked to user applications.

Wakaama became difficult to build and to maintain. Among other things,
there are defines and compiler flags that have to be set, the right set
of source files need to be selected and the correct include directories
have to be provided.

Therefore we rely on CMake and provide the users of Wakaama a static
library, which is relatively simple to build with a given set of
requested features.

This also helps with modularizing Wakaama to improve maintainability.