ci: Move to clang-format version 18
This makes use of the fact that we have moved CI to Ubuntu 24.04, and
are no longer constrained to the old versions offered by Ubuntu 20.04.

Please note: clang-format-18 does return an non-zero exit code when
changes have been found. For this reason, it gets now run in a subshell
which has the pipefail option disabled.